[refactor]split team module
이슈 번호 : #45
