+++ lib/itsm_web/live/common_k_create_vm_live/form.ex
@@ -7,6 +7,7 @@ defmodule ItsmWeb.CommonKCreateVmLive.Form do
   alias Itsm.Requests
   alias Itsm.Service.Request
   alias Itsm.Team
+  alias Itsm.Crews
 
   def mount(params, _session, socket) do
     crew_options = Accounts.crew_ids_names(socket.assigns.current_user)
